Simplify generated code for `derive(Display)`.
The fix in commit 2c74515 "Avoid `write!(f, "{}", #fmt)` in
`derive(Display)`" was very roundabout in that it:
- creates a closure that formats the fields,
- wraps it in a struct that impls Display by calling the closure,
- calls `Display::fmt()` on that struct.

This has been simplified to just formatting the fields. The wrapper
struct now *only* is actually used in the one case where it is needed:
when an enum has an outer `#[display("prefix {} suffix")]` attribute
with exactly one placeholder.