Split crate into a proc-macro and a regular library crate (#214)
proc-macro crates are loaded by the compiler instead of by the crate
that depends on it. This means it's not possible to use anything else
than exported anything else than exported proc-macros. But in some cases
we want the generated derives to use structs or traits that are provided
by the library. Two examples of this are:
1. An error type that can be used by the `Into` derive ([#173])
2. Or to support trait objects for the `source` of the `Error` derive
([#122])

This change splits the layout of this crate in two. One crate that only
exports proc-macros and is loaded by the compiler. And a second crate
that re-exports all these macros but is also able to export different
things such as structs and traits.

NOTE: This only refactors the code in this repo. No functional changes
happen in this change. These will be done in follow-up PRs.

[#173]: https://github.com/JelteF/derive_more/issues/173
[#122]: https://github.com/JelteF/derive_more/issues/122

Required for #173 #122

Co-authored-by: tyranron <tyranron@gmail.com>