Disable all features by default (#234, #210)
As discussed in #210 we want to disable all features by default in 1.0. With the amount of derives that are part of this library, the compile time of the library itself becomes significant. This way, we don't have to worry about increasing people their compile time when adding more derives.

On my machine, compiling the `derive_more-impl` crate took ~4.5 seconds when enabling all features. But when only enabling a single derive, it took 1.5 seconds.

Enabling all derives is still easy to do, by using the `full` feature.