Support Error in no_std environments (#268)
Resolves #261 

## Synopsis

The `Error` derive can be made to work well for the most part in
`no_std` environments by enabling `#![feature(error_in_core)]`. This
changes the `Error` derive slightly to import `Error` and related
traits from core, when the `std` feature is disabled.

In passing this also fixes actually running the nightly error tests.
They were not actually run anymore because there was no `build.rs` file
in the root of the repo, only in the `impl` package. So the `nightly`
config was not available in tests.

Co-authored-by: tyranron <tyranron@gmail.com>