Refactor `::derive_more` absolute paths to just `derive_more` (#344, #338)
## Synopsis

At the moment, the `derive_more` macros cannot be re-exported or the
`derive_more` crate cannot be renamed as the dependency, because
expansions require `::derive_more` module be present in the scope. See
#338 for details.




## Solution

Use `derive_more::*` paths instead of `::derive_more::*` in expansions,
allowing to provide this module in the scope on the call site even when
`derive_more` is transitive dependency of the crate.