Work around bug in rust-analyzer to make auto-imports work nicely (#376)
Resolves #351


## Synopsis

Rust-analyzer doesn't ignore `#[doc(hidden)]` exports when they use a
glob
pattern, only when they explicitely re-export. 


## Solution

This adds more level of indirection and one more glob pattern so that we
put
`#[doc(hidden)]` on explicit re-exports. This definitely starts to feel
like
we're doing hack upon hack here. But ah well, this seems to work. And
working
well with the most popular LSP for Rust seems worth a bit of a hack.