Replace `Self` type in `Into`/`TryInto` derives (#489, #413, #335)
## Synopsis

Resolves `Self` to the actual type it refers to in `Into`/`TryInto`
derives.
`Into`/`TryInto` derives generate `From`/`TryFrom` implementations on
the inner type instead of the enum/struct type on which the `#derive`
annotation is placed. During code generation, `Self`, which refers to
the enum/struct type, is copied verbatim from the enum/struct definition
to the `From`/`TryFrom` implementations where it unintendedly resolves
to the inner type.

## Solution

Use `syn::visit_mut` to substitute all instances 
of the `Self` type in `DeriveInput` with the
actual enum/struct type before code gen.

Based on:
<https://github.com/mbrobbel/narrow/blob/aaae3ddeda3c74e8a15324a5c853b898cf7f5c52/narrow-derive/src/util/self_replace.rs>

Co-authored-by: Kai Ren <tyranron@gmail.com>