Support `skip` attribute in `Eq`/`PartialEq` derives (#483, #163)
## Synopsis

Support of `skip` attribute on fields and variants:
```rust
#[derive(Eq, PartialEq)]
enum E {
    Foo(bool, #[eq(skip)] NoEq),
    #[partial_eq(skip)]
    Bar(NoEq, NoEq),
}
```

## Solution

Allow `utils::attr::Skip` on fields and variants.

### Ergonomics

`#[eq(skip)]`/`#[partial_eq(skip)]` attributes are visible and mean the
same for both `Eq` and `PartialEq` macros, allowing to use only one and
do not repeat yourself.

Once put on a variant (or a whole struct), ignores all the fields.