Support `Deref` and `DerefMut` derives for enums (#485, #388)
## Synopsis

Add `Deref` and `DerefMut` derive for enums. Only works on a single
field of each variant of the enum and as long as the field types are
compatible.
```rust
#[derive(Deref)]
enum NumRefEnum3 {
    Variant1 {
        #[deref(forward)]
        num: Box<i32>,
        useless: bool,
    },
    Variant2 {
        num: i32,
        #[deref(ignore)]
        useless: bool,
    },
    Variant3(i32),
}
```
See added docs and tests for more usage examples.

## Solution

Make sure each of the enum variants contains 1 and only 1 enabled field,
then generate a `match` block to extract the field. Checking if the
field types are compatible are unnecessary since this case will fail
compilation anyway.