Fix a non-working (and non-failing) test
My changes have broken a test, which actually was actually incorrect
twice: it wasn't doing the right thing (forwarding the From impl to
the field type), but it wasn't testing the right thing either so it
succeeded.

Make the test actually check that from forwarding works, and make the
forwarding actually work.